extract score input component
